Fix format string warnings when sizeof(suseconds_t) != sizeof(int)

+++ src/init.c
@@ -357,8 +357,8 @@ _dispatch_logv_init(void *context DISPATCH_UNUSED)
 			struct timeval tv;
 			gettimeofday(&tv, NULL);
 			fprintf(dispatch_logfile, "=== log file opened for %s[%u] at "
-					"%ld.%06u ===\n", getprogname() ?: "", getpid(),
-					tv.tv_sec, tv.tv_usec);
+					"%ld.%06ld ===\n", getprogname() ?: "", getpid(),
+					tv.tv_sec, (long)tv.tv_usec);
 			fflush(dispatch_logfile);
 		}
 	}
